oci: only push in parallel when forking
